修正mac机器下插件无法修正引擎问题

+++ extensions/fix_engine/src/main.ts
@@ -9,6 +9,12 @@ interface ConfigData{
 }
 export class _Helper {
 
+    private _isMac = false;
+    /**@description 是否是Mac平台 */
+    private get isMac(){
+        return this._isMac;
+    }
+
     /**@description creator 版本号 */
     private get appVersion(){
         return Editor.App.version;
@@ -21,9 +27,14 @@ export class _Helper {
             return this._path;
         }
         this._path = Editor.App.path;
-        //D:\Creator\Creator\3.1.0\resources\app.asar
+        //windows :  D:\Creator\Creator\3.1.0\resources\app.asar
+        //mac : Applications/CocosCreator/Creator/2.4.3/CocosCreator.app/Contents/MacOS --path
         let parser = path.parse(this._path);
         this._path = parser.dir;
+        if ( this._path.indexOf("/MacOS") != -1 ){
+            this._isMac = true;
+        }
+        this._path = this._path.replace("/MacOS","");
         return this._path;
     }
 
@@ -112,7 +123,6 @@ export class _Helper {
         console.log("Creator 安装路径 : " + this.appPath);
         console.log("Creator 引擎路径 : " + this.engineRoot);
         
-        
         let keys = Object.keys(this.config);
         for( let i = 0 ; i < keys.length ; i++ ){
             let data : ConfigData = this.config[keys[i]];
@@ -132,39 +142,54 @@ export class _Helper {
                 let sourcePath = `${path.join(__dirname,`../engine/${data.name}`)}`;
                 sourcePath = path.normalize(sourcePath);
                 let sourceData = fs.readFileSync(sourcePath,"utf-8");
-                let destData = fs.readFileSync(destPath,"utf-8");
-                let replace = function(){
-                    return arguments[1] + sourceData + arguments[3];
+                if ( fs.existsSync(destPath)){
+                    let destData = fs.readFileSync(destPath,"utf-8");
+                    let replace = function(){
+                        return arguments[1] + sourceData + arguments[3];
+                    }
+                    destData = destData.replace(/(declare\s*module\s*"cc"\s*\{)([\s\n\S]*)(export\s*function\s* murmurhash2_32_gc)/g,replace);
+                    fs.writeFileSync(destPath,destData,{encoding:"utf-8"});
+                    console.log(data.desc);
+                }else{
+                    console.error(`找不到引擎目录下文件:${destPath}`);
                 }
-                destData = destData.replace(/(declare\s*module\s*"cc"\s*\{)([\s\n\S]*)(export\s*function\s* murmurhash2_32_gc)/g,replace);
-                fs.writeFileSync(destPath,destData,{encoding:"utf-8"});
-                console.log(data.desc);
+                
             } else if( data.name == "jsbdts") {
                 //更新热更新声明文件
                 //(export\s*class\s*Manifest)([\s\n\S]*)(constructor\s*\(manifestUrl:\s*string\))
                 let destPath = `${this.appPath}/${data.path}`;
                 destPath = path.normalize(destPath);
-                let destData = fs.readFileSync(destPath,"utf-8");
-                let replaceManifest = function(){
-                    return arguments[1] + HotUpdateDTS.manifest + arguments[3];
+                if ( fs.existsSync(destPath)){
+                    let destData = fs.readFileSync(destPath,"utf-8");
+                    let replaceManifest = function(){
+                        return arguments[1] + HotUpdateDTS.manifest + arguments[3];
+                    }
+                    destData = destData.replace(/(export\s*class\s*Manifest\s*\{)([\s\n\S]*)(constructor\s*\(manifestUrl:\s*string\))/g,replaceManifest);
+                    let replaceAssetsManager = function(){
+                        return arguments[1] + HotUpdateDTS.assetsManager + arguments[3];
+                    }
+                    destData = destData.replace(/(export\s*class\s*AssetsManager\s*\{)([\s\n\S]*)(constructor\s*\(manifestUrl:\s*string)/g,replaceAssetsManager);
+                    fs.writeFileSync(destPath,destData,{encoding:"utf-8"});
+                    console.log(data.desc);
+                }else{
+                    console.error(`找不到引擎目录下文件:${destPath}`);
                 }
-                destData = destData.replace(/(export\s*class\s*Manifest\s*\{)([\s\n\S]*)(constructor\s*\(manifestUrl:\s*string\))/g,replaceManifest);
-                let replaceAssetsManager = function(){
-                    return arguments[1] + HotUpdateDTS.assetsManager + arguments[3];
-                }
-                destData = destData.replace(/(export\s*class\s*AssetsManager\s*\{)([\s\n\S]*)(constructor\s*\(manifestUrl:\s*string)/g,replaceAssetsManager);
-                fs.writeFileSync(destPath,destData,{encoding:"utf-8"});
-                console.log(data.desc);
             } else{
                 //查看本地是否有文件
                 let sourcePath = `${path.join(__dirname,`../engine/${data.name}`)}`;
                 sourcePath = path.normalize(sourcePath);
                 let destPath = `${this.engineRoot}/${data.path}`;
                 destPath = path.normalize(destPath);
-                if( fs.existsSync(sourcePath) ){
-                    let sourceData = fs.readFileSync(sourcePath,"utf-8");
-                    fs.writeFileSync(destPath,sourceData,{encoding:"utf-8"});
-                    console.log(data.desc);
+                if ( fs.existsSync(destPath) ){
+                    if( fs.existsSync(sourcePath) ){
+                        let sourceData = fs.readFileSync(sourcePath,"utf-8");
+                        fs.writeFileSync(destPath,sourceData,{encoding:"utf-8"});
+                        console.log(data.desc);
+                    }else{
+                        console.error(`找不到源文件:${sourcePath}`);
+                    }
+                }else{
+                    console.error(`找不到引擎目录下文件:${destPath}`);
                 }
             }
         }
